Change cFE build system to compile with musl
Composite currently builds with musl libc.  Before this, cFE compiled with glibc.
This caused a number of problems, including structs that needed to be passed back and
forth, which were defined with different layouts in each libc.
Also, for consistency, we wanted cFE and composite to be compiled under the same libc.
The cFE is also designed for use on embedded systems so glibc is inappropriate.