Implement a read / write filesystem that satisifies all osal requirements
The filesystem now passes all of the pertinent unit tests.  It implements
opened files and directory streams with file descriptors in a sensable way
it supports permissions, as well as statically linked files from
build time, as well as reading and writting to new files.  It follows
similar function calls to posix filesystems.